feat: basic menu
I've added a basic menu to the title screen that does absolutely nothing whatsoever. But it does implement the Button class as a basic example. In other news while work for a better deltaTime is worked on, a delta int variable is being used to ensure time passes correctly according to the frame rate